Move cbor decode functionality into its own file
