+++ packages/sdk/src/plt/Cbor.ts
@@ -2,15 +2,7 @@ import { Buffer } from 'buffer/index.js';
 
 import * as Proto from '../grpc-api/v2/concordium/protocol-level-tokens.js';
 import { HexString } from '../types.js';
-import { cborDecode, cborEncode } from '../types/cbor.js';
-import {
-    TokenAmount,
-    TokenHolder,
-    TokenListUpdateEventDetails,
-    TokenMetadataUrl,
-    TokenPauseEventDetails,
-} from './index.js';
-import { TokenInitializationParameters, TokenModuleAccountState, TokenModuleState } from './module.js';
+import { cborEncode } from '../types/cbor.js';
 
 export type JSON = HexString;
 
@@ -121,158 +113,7 @@ export function toProto(cbor: Cbor): Proto.CBor {
     };
 }
 
-function decodeTokenModuleState(value: Cbor): TokenModuleState {
-    const decoded = cborDecode(value.bytes);
-    if (typeof decoded !== 'object' || decoded === null) {
-        throw new Error('Invalid CBOR data for TokenModuleState');
-    }
-
-    // Validate required fields
-    if (!('governanceAccount' in decoded && TokenHolder.instanceOf(decoded.governanceAccount))) {
-        throw new Error('Invalid TokenModuleState: missing or invalid governanceAccount');
-    }
-    if (!('metadata' in decoded)) {
-        throw new Error('Invalid TokenModuleState: missing metadataUrl');
-    }
-    let metadata = TokenMetadataUrl.fromCBORValue(decoded.metadata);
-    if (!('name' in decoded && typeof decoded.name === 'string')) {
-        throw new Error('Invalid TokenModuleState: missing or invalid name');
-    }
-
-    // Validate optional fields
-    if ('allowList' in decoded && typeof decoded.allowList !== 'boolean') {
-        throw new Error('Invalid TokenModuleState: allowList must be a boolean');
-    }
-    if ('denyList' in decoded && typeof decoded.denyList !== 'boolean') {
-        throw Error('Invalid TokenModuleState: denyList must be a boolean');
-    }
-    if ('mintable' in decoded && typeof decoded.mintable !== 'boolean') {
-        throw new Error('Invalid TokenModuleState: mintable must be a boolean');
-    }
-    if ('burnable' in decoded && typeof decoded.burnable !== 'boolean') {
-        throw new Error('Invalid TokenModuleState: burnable must be a boolean');
-    }
-    if ('paused' in decoded && typeof decoded.paused !== 'boolean') {
-        throw new Error('Invalid TokenModuleState: paused must be a boolean');
-    }
-
-    return { ...decoded, metadata } as TokenModuleState;
-}
-
-function decodeTokenModuleAccountState(value: Cbor): TokenModuleAccountState {
-    const decoded = cborDecode(value.bytes);
-    if (typeof decoded !== 'object' || decoded === null) {
-        throw new Error('Invalid CBOR data for TokenModuleAccountState');
-    }
-
-    // Validate optional fields
-    if ('allowList' in decoded && typeof decoded.allowList !== 'boolean') {
-        throw new Error('Invalid TokenModuleState: allowList must be a boolean');
-    }
-    if ('denyList' in decoded && typeof decoded.denyList !== 'boolean') {
-        throw Error('Invalid TokenModuleState: denyList must be a boolean');
-    }
-
-    return decoded as TokenModuleAccountState;
-}
-
-function decodeTokenListUpdateEventDetails(value: Cbor): TokenListUpdateEventDetails {
-    const decoded = cborDecode(value.bytes);
-    if (typeof decoded !== 'object' || decoded === null) {
-        throw new Error(`Invalid event details: ${JSON.stringify(decoded)}. Expected an object.`);
-    }
-    if (!('target' in decoded && TokenHolder.instanceOf(decoded.target))) {
-        throw new Error(`Invalid event details: ${JSON.stringify(decoded)}. Expected 'target' to be a TokenHolder`);
-    }
-
-    return decoded as TokenListUpdateEventDetails;
-}
-
-function decodeTokenPauseEventDetails(value: Cbor): TokenPauseEventDetails {
-    const decoded = cborDecode(value.bytes);
-    if (typeof decoded !== 'object' || decoded === null) {
-        throw new Error(`Invalid event details: ${JSON.stringify(decoded)}. Expected an object.`);
-    }
-
-    return decoded as TokenPauseEventDetails;
-}
-
-function decodeTokenInitializationParameters(value: Cbor): TokenInitializationParameters {
-    const decoded = cborDecode(value.bytes);
-    if (typeof decoded !== 'object' || decoded === null) {
-        throw new Error('Invalid CBOR data for TokenInitializationParameters');
-    }
-
-    // Validate required fields
-    if (!('governanceAccount' in decoded && TokenHolder.instanceOf(decoded.governanceAccount))) {
-        throw new Error('Invalid TokenInitializationParameters: missing or invalid governanceAccount');
-    }
-    if (!('metadata' in decoded)) {
-        throw new Error('Invalid TokenInitializationParameters: missing metadataUrl');
-    }
-    let metadata = TokenMetadataUrl.fromCBORValue(decoded.metadata);
-    if (!('name' in decoded && typeof decoded.name === 'string')) {
-        throw new Error('Invalid TokenInitializationParameters: missing or invalid name');
-    }
-
-    // Validate optional fields
-    if ('allowList' in decoded && typeof decoded.allowList !== 'boolean') {
-        throw new Error('Invalid TokenInitializationParameters: allowList must be a boolean');
-    }
-    if ('denyList' in decoded && typeof decoded.denyList !== 'boolean') {
-        throw Error('Invalid TokenInitializationParameters: denyList must be a boolean');
-    }
-    if ('mintable' in decoded && typeof decoded.mintable !== 'boolean') {
-        throw new Error('Invalid TokenInitializationParameters: mintable must be a boolean');
-    }
-    if ('burnable' in decoded && typeof decoded.burnable !== 'boolean') {
-        throw new Error('Invalid TokenInitializationParameters: burnable must be a boolean');
-    }
-    if ('paused' in decoded && typeof decoded.paused !== 'boolean') {
-        throw new Error('Invalid TokenInitializationParameters: paused must be a boolean');
-    }
-
-    // Optional initial supply
-    if ('initialSupply' in decoded && !TokenAmount.instanceOf(decoded.initialSupply)) {
-        throw new Error(`Invalid TokenInitializationParameters: Expected 'initialSupply' to be of type 'TokenAmount'`);
-    }
-
-    return { ...decoded, metadata } as TokenInitializationParameters;
-}
-
-type DecodeTypeMap = {
-    TokenModuleState: TokenModuleState;
-    TokenModuleAccountState: TokenModuleAccountState;
-    TokenListUpdateEventDetails: TokenListUpdateEventDetails;
-    TokenPauseEventDetails: TokenPauseEventDetails;
-    TokenInitializationParameters: TokenInitializationParameters;
-};
-
-export function decode<T extends keyof DecodeTypeMap>(cbor: Cbor, type: T): DecodeTypeMap[T];
-export function decode(cbor: Cbor, type?: undefined): unknown;
-
-/**
- * Decode CBOR encoded data into its original representation.
- * @param {Cbor} cbor - The CBOR encoded data.
- * @param {string | undefined} type - Optional type hint for decoding.
- * @returns {unknown} The decoded data.
- */
-export function decode<T extends keyof DecodeTypeMap | undefined>(cbor: Cbor, type: T): unknown {
-    switch (type) {
-        case 'TokenModuleState':
-            return decodeTokenModuleState(cbor);
-        case 'TokenModuleAccountState':
-            return decodeTokenModuleAccountState(cbor);
-        case 'TokenListUpdateEventDetails':
-            return decodeTokenListUpdateEventDetails(cbor);
-        case 'TokenPauseEventDetails':
-            return decodeTokenPauseEventDetails(cbor);
-        case 'TokenInitializationParameters':
-            return decodeTokenInitializationParameters(cbor);
-        default:
-            return cborDecode(cbor.bytes);
-    }
-}
+export { decode } from './decode.js';
 
 /**
  * Encode a value into CBOR format.
